runtime.c -> libflipper.c (merge)
